winget-source: Fix BIND_ADDRESS in Nodejs 18+
When accessing a domain, nodejs would only use the first resolved IP (either v6 or v4) in onlookup(). When a IPv4 address is binded and the first DNS result is IPv6 address, internalConnect() would throw a bind error (as it calls libuv's IPv6-related function).

By giving family to globalAgent this bug is workarounded.